+++ safire/include/safire/safire.h
@@ -8,19 +8,12 @@ extern "C" {
 // main includes
 #include "pipeline.h"
 #include "ecs.h"
+#include "components.h"
 #include "input.h"
 #include "utils.h"
 
-typedef struct SFRcontext               SFRcontext_t;
-typedef struct SFRcreate_info           SFRcreate_info_t;
 
 
-
-
-struct SFRcreate_info {
-  const char*               window_title;
-  uint32_t                  window_size[2];
-};
 
 SAFIRE_API void             safire(const char* window_title, int window_width, int window_height, 
                                    bool fullscreen, SFRscene_t** scenes, uint32_t scenes_count);
@@ -31,6 +24,9 @@ SAFIRE_API float            sfr_elapsed_time();
 
 SAFIRE_API SFRentity_t*     sfr_instantiate(const char* name);
 SAFIRE_API void             sfr_destroy(SFRentity_t* entity);
+
+
+
 
 #ifdef __cplusplus
 }