+++ topdown_shooter/include/topdown_shooter/manager.h
@@ -1,8 +1,0 @@
-#ifndef __TDS_MANAGER_H__
-#define __TDS_MANAGER_H__
-
-#include <safire/utils.h>
-
-SAFIRE_API void print_message_from_manager();
-
-#endif // __TDS_MANAGER_H__