+++ topdown_shooter/src/manager.c
@@ -1,5 +1,0 @@
-#include <topdown_shooter/manager.h>
-
-void print_message_from_manager() {
-  printf("this is a message from the manger in topdown_shooter\n");
-}