+++ safire/CHANGELOG.md
@@ -4,8 +4,13 @@
 
 * made code more cleaner *(VERSION 0.0.6.1)*
   * spacing and tabbing for functions and type defines in the header files
+* updated README
+  * instructions for building program
 * ecs implementation
   * finding entities and components *(VERSION 0.0.6.2)*
+  * create entities *(VERSION 0.0.7)*
+  * create and add components to entities
+  * erase entity and its components from the scene
 
 ## VERSION 0.0.6 (25/4/2022)